print reqs info after installation msg
